Fixed issue with outflow calculation. Previously, the `AddOutflow` method would be called *only* if at a boundary face *AND* and also not at a reflecting boundary face. This was incorrect, as outflow physically occurs whether the boundary is vacuum, isotropic, or reflecting.
Previously, due to the incorrect check, the `hdpe_balance.py` test failed as the calculated final in- and out-flow rates were 0. This is incorrect, however, as *some* of the particles hit the reflecting boundary faces and reflect back into the domain, so the rates are non-zero. After this corrected check, the test passes.